backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
python-pptx==0.6.23
pytesseract==0.3.13
Pillow==10.4.0
